+++ cmp/dataview/DataViewModel.ts
@@ -29,12 +29,16 @@ import {
 import {bindable, makeObservable} from '@xh/hoist/mobx';
 import {throwIf} from '@xh/hoist/utils/js';
 import {isFunction, isNumber} from 'lodash';
-import {ReactNode} from 'react';
 
 /**
  * Configuration for a DataView.
+ *
+ *  Additional properties not specified here will be passed to the underlying
+ *  GridModel. Note this is for advanced usage - not all configs supported, and many will
+ *  override DataView defaults in ways that will break this component.
  */
-export interface DataViewConfig {
+export interface DataViewConfig extends GridConfig {
+    // TODO: Accept grid keys without publicizing them?
     /** A Store instance, or a config to create one. */
     store?: Store | StoreConfig;
 
@@ -70,7 +74,7 @@ export interface DataViewConfig {
     selModel?: StoreSelectionModel | StoreSelectionConfig | 'single' | 'multiple' | 'disabled';
 
     /** Text/HTML to display if view has no records.*/
-    emptyText?: ReactNode;
+    emptyText?: string;
 
     /** True (default) to hide empty text until after the Store has been loaded at least once. */
     hideEmptyTextBeforeLoad?: boolean;
@@ -114,13 +118,6 @@ export interface DataViewConfig {
      * the row's data. (Note that this may be null - e.g. for clicks on full-width group rows.)
      */
     onRowDoubleClicked?: (e: any) => void;
-
-    /**
-     * "Escape hatch" object to pass directly to GridModel. Note these options may be used
-     * / overwritten by the framework itself, and are not all guaranteed to be compatible
-     * with its usages of GridModel.
-     */
-    gridOptions?: Omit<GridConfig, keyof DataViewConfig>;
 }
 
 export type ItemHeightFn = (params: {
@@ -164,7 +161,7 @@ export class DataViewModel extends HoistModel {
             rowClassRules,
             onRowClicked,
             onRowDoubleClicked,
-            gridOptions
+            ...restArgs
         } = config;
 
         throwIf(
@@ -208,7 +205,7 @@ export class DataViewModel extends HoistModel {
             onRowClicked,
             onRowDoubleClicked,
             columns,
-            ...gridOptions
+            ...restArgs
         });
     }