+++ utils/datetime/DateTimeUtils.ts
@@ -4,7 +4,6 @@
  *
  * Copyright © 2023 Extremely Heavy Industries Inc.
  */
-import {isDate, isNil} from 'lodash';
 import moment from 'moment';
 
 export const MILLISECONDS = 1,
@@ -22,9 +21,7 @@ export const MILLISECONDS = 1,
  * Returns true if input time is null.
  */
 export function olderThan(t: Date | number, millis: number) {
-    if (isNil(t)) return true;
-    t = isDate(t) ? t.getTime() : t;
-    return t < Date.now() - millis;
+    return !t || t < Date.now() - millis;
 }
 
 /** Returns midnight browser local time for the provided date, or current date. */